don't just delete tag but remove tagged-items
Mainly fixed an issue with tag deletion. The api must not just delete a tag, but remove it from all items of the currently relevant model. (Since one tag could be used for multiple different models.) The tag will be only deleted if no related tagged-item is left.

+++ taggit_ui/views.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 
 from django.http import Http404
+from django.contrib.contenttypes.models import ContentType
 from rest_framework import status
 from rest_framework.permissions import BasePermission, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from taggit.models import Tag
+from taggit.models import TaggedItem
 
 
 class CanDeleteTag(BasePermission):
@@ -13,14 +15,24 @@ class CanDeleteTag(BasePermission):
         return request.user.has_perm('taggit.delete_tag')
 
 
-class DeleteTagAPI(APIView):
+class RemoveTagApi(APIView):
     permission_classes = [IsAuthenticated, CanDeleteTag]
 
-    def delete(self, request, id=None):
+    def delete(self, request, app_label=None, model_name=None, tag_id=None):
         try:
-            tag = Tag.objects.get(pk=id)
+            tag = Tag.objects.get(pk=tag_id)
         except Tag.DoesNotExist:
-            raise Http404
-        else:
+            raise Http404('Tag not found: {}'.format(tag_id))
+
+        try:
+            content_type = ContentType.objects.get(app_label=app_label, model=model_name)
+        except ContentType.DoesNotExist:
+            raise Http404('Model not found: {}.{}'.format(app_label, model_name))
+
+        TaggedItem.objects.filter(content_type=content_type, tag=tag).delete()
+
+        # Cleanup tag if no tagged-items left.
+        if not TaggedItem.objects.filter(tag=tag):
             tag.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+        
+        return Response(status=status.HTTP_204_NO_CONTENT)
